requirements.txt
alabaster==0.7.12
aniso8601==9.0.1
asgiref==3.3.4
Babel==2.9.0
beautifulsoup4==4.9.3
boto3==1.17.53
botocore==1.20.53
Brotli==1.0.9
calmjs.parse==1.2.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
csscompressor==0.9.5
cssselect==1.1.0
cycler==0.10.0
dj-database-url==0.5.0
Django==3.2
django-appconf==1.0.4
django-bmemcached==0.3.0
django-bootstrap4==3.0.0
django-ckeditor==6.0.0
django-cleanup==5.2.0
django-compression-middleware==0.4.1
django-compressor==2.4.1
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-htmlmin==0.11.0
django-js-asset==1.2.2
django-pwa==1.0.10
django-s3-folder-storage==0.5
django-storages==1.11.1
docutils==0.16
Flask==1.1.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.8
fontawesome==5.10.1.post1
futures==3.1.1
gunicorn==20.1.0
h5py==3.2.1
html5lib==1.1
idna==2.10
imagesize==1.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
keras==2.6.0
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
markdown2==2.4.0
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.20.2
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
pip-check-reqs==2.3.0
ply==3.11
psycopg2==2.8.6
psycopg2-binary==2.8.6
Pygments==2.8.1
pylibmc==1.6.1
pyparsing==2.4.7
pyperclip==1.8.2
pyquery==1.4.3
python-binary-memcached==0.30.1
python-dateutil==2.8.1
python-memcached==1.59
pytz==2021.1
pyuploadcare==2.7.0
PyYAML==5.4.1
rcssmin==1.0.6
readtime==1.1.1
requests==2.25.1
rjsmin==1.1.0
s3==3.0.0
s3transfer==0.3.7
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==3.5.4
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.60.0
transfer==0.40
uhashring==2.1
urllib3==1.26.4
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
xmltodict==0.12.0
zstandard==0.15.2
